Refactor to prepare for follow-on change
